+++ core/src/main/java/org/seedstack/seed/core/internal/validation/MethodValidationInterceptor.java
@@ -10,7 +10,6 @@ package org.seedstack.seed.core.internal.validation;
 import org.aopalliance.intercept.MethodInterceptor;
 import org.aopalliance.intercept.MethodInvocation;
 import org.seedstack.seed.SeedException;
-import org.seedstack.seed.core.VerboseConstraintViolationException;
 
 import javax.validation.ConstraintViolation;
 import javax.validation.ValidatorFactory;