+++ specs/src/main/java/org/seedstack/seed/spi/ToolContext.java
@@ -1,0 +1,26 @@
+/**
+ * Copyright (c) 2013-2016, The SeedStack authors <http://seedstack.org>
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package org.seedstack.seed.spi;
+
+/**
+ * Describes a tool execution context.
+ *
+ * @author adrien.lauer@mpsa.com
+ */
+public class ToolContext {
+
+    private final String[] args;
+
+    public ToolContext(String[] args) {
+        this.args = args;
+    }
+
+    public String[] getArgs() {
+        return args;
+    }
+}